Allow num0 to float

+++ calc.py
@@ -3,7 +3,7 @@
 # '1 + 2' -> 3, or '1 - 2' -> -1.
 def compute(expression):
     num0, operator, num1 = expression.split(' ')
-    num0, num1 = int(num0), int(num1)
+    num0, num1 = float(num0), int(num1)
     if operator == '+':
         return num0 + num1
     elif operator == '-':
